renamed Extensions.kt -> Conversions.kt and separated them by layers
