+++ app/src/main/kotlin/com/igor_shaula/api_polling/data_layer/VehiclesRepositoryContract.kt
@@ -2,6 +2,7 @@ package com.igor_shaula.api_polling.data_layer
 
 import androidx.lifecycle.MutableLiveData
 
+// for now this thing exists for only description of interface for the consumer (as ViewModel)
 interface VehiclesRepositoryContract {
 
     val mainDataStorage: MutableLiveData<MutableMap<String, VehicleRecord>>
