+++ app/src/main/kotlin/com/igor_shaula/api_polling/data_layer/network_data_source/NetworkDataSource.kt
@@ -3,7 +3,6 @@ package com.igor_shaula.api_polling.data_layer.network_data_source
 import com.igor_shaula.api_polling.data_layer.VehicleDetailsRecord
 import com.igor_shaula.api_polling.data_layer.VehicleRecord
 import com.igor_shaula.api_polling.data_layer.network_data_source.retrofit.VehicleRetrofitNetworkServiceImpl
-import com.igor_shaula.api_polling.data_layer.toVehicleItemRecords
 
 class NetworkDataSource {
